environment_setup/requirements.txt
appnope==0.1.4
asttokens==2.4.1
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.3.1
gmpy2==2.1.2
idna==3.4
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.5.0
kiwisolver==1.4.5
lightning-utilities==0.11.2
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.1
numpy==1.26.4
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wcwidth==0.2.13
wheel==0.41.2
zipp==3.17.0
